+++ install_condocam.ai.sh
@@ -46,14 +46,18 @@ else
   SD_CARD_PATH="/dev/mmcblk0"
 fi
 
-RASPI_OS_TYPE="lite" # or "desktop"
-RASPI_OS_ID="raspios_armhf" #or "raspios_lite" 
-if [[ "${RASPI_OS_TYPE}" == "lite" ]]; then
-	RASPI_OS_ID="raspios_lite"
-fi
+RPI_IMAGE_URL="https://downloads.raspberrypi.org/raspios_lite_arm64/images/raspios_lite_arm64-2022-04-07/2022-04-04-raspios-bullseye-arm64-lite.img.xz"
+USE_LATEST_RASPI_OS=false
+
+RASPI_OS_TYPE="lite" # or "full"
+RASPI_CPU_TYPE="arm64" # or "armhf
+
+RASPI_OS_ID="raspios_${RASPI_OS_TYPE}_${RASPI_CPU_TYPE}"
 
 #get HOSTNAME for raspberry pi from firstrun.sh
 RPI_HOST_NAME=$( grep "^HOSTNAME=" condocam-pi/firstrun.sh | cut -d "=" -f 2 )
+#get USERNAME for raspberry pi from firstrun.sh
+RPI_USER_NAME=$( grep "^USERNAME=" condocam-pi/firstrun.sh | cut -d "=" -f 2 )
 
 echo "SD_CARD_PATH = ${SD_CARD_PATH}"
 echo "RPI_HOST_NAME = ${RPI_HOST_NAME}"
@@ -101,43 +105,49 @@ if [[ ! "${REPLY}" =~ ^[Yy]$ ]]; then
     exit 1
 fi 
 
-echo 
-echo "=============================================================="
-echo " get latest raspbian 32bit os image"
-echo "=============================================================="
-echo 
+echo "USE_LATEST_RASPI_OS==${USE_LATEST_RASPI_OS}"
+if [[ ${USE_LATEST_RASPI_OS} == true ]] ; then
+  echo 
+  echo "=============================================================="
+  echo " get URL of latest Raspberry Pi OS 64bit image"
+  echo "=============================================================="
+  echo 
 
-echo "get rasbian os information data from server";
-rm operating-systems-categories.json
-wget https://downloads.raspberrypi.org/operating-systems-categories.json
+  echo "get rasbian os information data from server";
+  rm operating-systems-categories.json
+  wget https://downloads.raspberrypi.org/operating-systems-categories.json
 
-RPI_IMAGE_URL=$( <operating-systems-categories.json grep "${RASPI_OS_ID}" | grep urlHttp | sed -e 's/.*\: \"\(.*\)\".*/\1/' )
+  RPI_IMAGE_URL=$( <operating-systems-categories.json grep "${RASPI_OS_ID}" | grep urlHttp | sed -e 's/.*\: \"\(.*\)\".*/\1/' )
+  echo "latest image URL is: ${RPI_IMAGE_URL}"
+fi
 RPI_IMAGE_HASH_URL="${RPI_IMAGE_URL}.sha256"
 
 echo 
 echo "=============================================================="
-echo " check downloaded file"
+echo " download and check Raspberry Pi OS"
 echo "=============================================================="
 echo 
 
-RPI_IMAGE_ZIP=$(basename "${RPI_IMAGE_URL}")
+RPI_IMAGE_XZ=$(basename "${RPI_IMAGE_URL}")
 RPI_IMAGE_HASH=$(basename "${RPI_IMAGE_HASH_URL}")
-RPI_IMAGE="${RPI_IMAGE_ZIP//.zip/.img}"
+RPI_IMAGE="${RPI_IMAGE_XZ//.xz/}"
 
-if [ -f "${RPI_IMAGE_ZIP}" ]; then
-  echo "${RPI_IMAGE_ZIP} file found. Skipping download."
+if [ -f "${RPI_IMAGE}" ]; then
+  echo "${RPI_IMAGE} file found. Skipping download."
+elif [ -f "${RPI_IMAGE_XZ}" ]; then
+  echo "${RPI_IMAGE_XZ} file found. Skipping download."
 else
   echo "downloading Raspberry Pi OS image from server. please wait ..."
-  rm   "${RPI_IMAGE_ZIP}.downloading"
-  wget "${RPI_IMAGE_URL}" -O "${RPI_IMAGE_ZIP}.downloading"
-  mv   "${RPI_IMAGE_ZIP}.downloading" "${RPI_IMAGE_ZIP}"
+  rm   "${RPI_IMAGE_XZ}.downloading"
+  wget "${RPI_IMAGE_URL}" -O "${RPI_IMAGE_XZ}.downloading"
+  mv   "${RPI_IMAGE_XZ}.downloading" "${RPI_IMAGE_XZ}"
 
   echo "downloading hash file for image file from server. please wait ..."
   rm "${RPI_IMAGE_HASH}"
   wget "${RPI_IMAGE_HASH_URL}"
 
   echo "checking hash value of image file"
-  HASH_OK=$( sha256sum -c "${RPI_IMAGE_HASH}" | grep "${RPI_IMAGE_ZIP}: OK" )
+  HASH_OK=$( sha256sum -c "${RPI_IMAGE_HASH}" | grep "${RPI_IMAGE_XZ}: OK" )
   if [ -z "${HASH_OK}" ]; then
     echo "hash does not match, aborting"
     exit
@@ -148,17 +158,17 @@ fi
 
 echo 
 echo "=============================================================="
-echo " unzip image file"
+echo " extract image file"
 echo "=============================================================="
 echo 
 
-echo "unzip the Raspberry Pi OS image"
+echo "extract the Raspberry Pi OS image"
 if [ -f "${RPI_IMAGE}" ]; then
-  echo "file found, skip the zip"
+  echo "file found, skip the extract"
 else
-  echo "unzipping file. please wait a few minutes ..."
-  echo "unzip -o ${RPI_IMAGE_ZIP} ${RPI_IMAGE}"
-  unzip -o "${RPI_IMAGE_ZIP}" "${RPI_IMAGE}"
+  echo "extracting file. please wait a few minutes ..."
+  echo "unxz ${RPI_IMAGE_XZ}"
+  unxz "${RPI_IMAGE_XZ}"
 fi
 
 echo 
@@ -168,8 +178,8 @@ echo "=============================================================="
 echo 
 
 echo "unmount SD card $DISKNAME"
-echo "press any key to continue..."
-read -rn 1 -s
+#echo "press any key to continue..."
+#read -rn 1 -s
 
 mount | grep "${DISKNAME}" | cut -d " " -f 3 | while read -r line ; do
     echo "sudo umount ${line}"
@@ -183,15 +193,15 @@ read -rn 1 -s
 echo
 
 echo "wipe SD card: sudo wipefs -a ${SD_CARD_PATH}"
-echo "press any key to continue..."
-read -rn 1 -s
+#echo "press any key to continue..."
+#read -rn 1 -s
 sudo wipefs -a "${SD_CARD_PATH}"
 echo
 
 echo "write image file to SD card: sudo dd bs=8M if=${RPI_IMAGE} of=${SD_CARD_PATH} status=progress"
-echo "press any key to continue..."
-read -rn 1 -s
-echo "writing image to SD card ..."
+#echo "press any key to continue..."
+#read -rn 1 -s
+#echo "writing image to SD card ..."
 sudo dd bs=8M if="${RPI_IMAGE}" of="${SD_CARD_PATH}" status=progress
 echo
 
@@ -208,9 +218,9 @@ echo "udisksctl mount -b \"/dev/disk/by-uuid/${UUID}\""
 udisksctl mount -b "/dev/disk/by-uuid/${UUID}"
 RPI_PATH=$( mount | grep "${DISKNAME}" | cut -d " " -f 3 )
 
-echo "press any key to continue..."
-read -rn 1 -s
-echo
+#echo "press any key to continue..."
+#read -rn 1 -s
+#echo
 
 echo 
 echo "=============================================================="
@@ -220,7 +230,7 @@ echo
 
 echo "the UUID of the root partition might have changed for the downloaded image. Updating the entry in cmdline.txt"
 PARTUUID=$( sudo blkid | grep rootfs | grep -oP '(?<=PARTUUID=\").*(?=\")' )
-echo "set PARTUUID=$PARTUUID for rootfs in cmdline.txt"
+echo "set PARTUUID=$PARTUUID for rootfs in condocam-pi/cmdline.txt"
 sed -i "s/\(.*PARTUUID=\)[^ ]*\( .*\)/\1$PARTUUID\2/" condocam-pi/cmdline.txt
 
 echo 
@@ -234,9 +244,9 @@ echo "cp -r condocam-pi/* ${RPI_PATH}"
 cp -r condocam-pi/* "${RPI_PATH}"
 echo
 
-echo "press any key to continue..."
-read -rn 1 -s
-echo
+#echo "press any key to continue..."
+#read -rn 1 -s
+#echo
 
 echo 
 echo "=============================================================="
@@ -264,8 +274,8 @@ echo " setup telegram bot!"
 echo "=============================================================="
 echo 
 
-echo "before you continue, send any message to your telegram bot!"
-echo "press any key to continue..."
+echo "BEFORE you continue, send any message to your telegram bot!"
+echo "Then press any key to continue..."
 read -rn 1 -s
 echo
 echo "all work is done. Please insert the SD card into your raspberry pi"
@@ -274,7 +284,7 @@ echo "NOTE: when starting up, your raspberry pi should reboot 4 times until all 
 echo
 echo "///////////////////////////////////////////////////////////////"
 echo
-echo "               ssh -x pi@${RPI_HOST_NAME}.local"
+echo "               ssh -x ${RPI_USER_NAME}@${RPI_HOST_NAME}.local"
 echo "               http://${RPI_HOST_NAME}:8765/"
 echo
 echo "///////////////////////////////////////////////////////////////"
